Convert GPX points to Polygons
Clean up file names, and remove unnecessary spaces or characters.
Convert 1000s of GPX to points to lines to polygons and then dissolve to one feature. Make sure it is located in right country. If not, write file to a list.
Try/except error handling - move bad data that cannot be converted to a bad data folder.